Move the helper function to a dedicated file.
