+++ front/lib/http_api/group_header.ts
@@ -1,0 +1,12 @@
+import { DustGroupIdsHeader } from "@dust-tt/types";
+
+export function getGroupIdsFromHeaders(
+  headers: Record<string, string | string[] | undefined>
+): string[] | undefined {
+  const groupIds = headers[DustGroupIdsHeader];
+  if (typeof groupIds === "string") {
+    return groupIds.split(",").map((id) => id.trim());
+  } else {
+    undefined;
+  }
+}
