+++ front/pages/api/v1/w/[wId]/assistant/conversations/[cId]/messages/index.ts
@@ -13,11 +13,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { getConversation } from "@app/lib/api/assistant/conversation";
 import { postUserMessageWithPubSub } from "@app/lib/api/assistant/pubsub";
-import {
-  Authenticator,
-  getAPIKey,
-  getGroupIdsFromHeaders,
-} from "@app/lib/auth";
+import { Authenticator, getAPIKey } from "@app/lib/auth";
+import { getGroupIdsFromHeaders } from "@app/lib/http_api/group_header";
 import { apiError, withLogging } from "@app/logger/withlogging";
 
 export type PostMessagesResponseBody = {
